- Extracted renderable/templatable classes
